% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plotchart_aws.R
\name{chartMinAWSData}
\alias{chartMinAWSData}
\title{Get 10 or 15 minutes data.}
\usage{
chartMinAWSData(net_aws, var_hgt, stat, start, end, plotrange, aws_dir)
}
\arguments{
\item{net_aws}{the network code and AWS ID, form <network code>_<AWS ID>. AWS network code, 1: campbell, 2: adcon}

\item{var_hgt}{the variable code and observation height, form  <var code>_<height>.}

\item{stat}{statistic code.}

\item{start}{start time.}

\item{end}{end time.}

\item{plotrange}{get range.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Get 10 or 15 minutes data to display on chart.
}
