% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rain_accumulation.R
\name{chartRainAccumul}
\alias{chartRainAccumul}
\title{Compute precipitation accumulation.}
\usage{
chartRainAccumul(tstep, net_aws, start, end, accumul, aws_dir)
}
\arguments{
\item{tstep}{time basis to accumulate the data.}

\item{net_aws}{a vector of the network code and AWS ID, form <network code>_<AWS ID>. 
AWS network code, 1: campbell, 2: adcon}

\item{start}{start date.}

\item{end}{end date.}

\item{accumul}{accumulation duration.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Compute precipitation accumulation for chart display.
}
