% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_display_winddata.R
\name{chartWindRose}
\alias{chartWindRose}
\title{Get wind data.}
\usage{
chartWindRose(net_aws, height, tstep, start, end, aws_dir)
}
\arguments{
\item{net_aws}{the network code and AWS ID, form <network code>_<AWS ID>. AWS network code, 1: campbell, 2: adcon}

\item{height}{the observation height.}

\item{tstep}{time step of the data.}

\item{start}{start time.}

\item{end}{end time.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Get wind data to display wind rose.
}
