% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downAWSAggrCDTDataCSV}
\alias{downAWSAggrCDTDataCSV}
\title{Get aggregated data in CDT format.}
\usage{
downAWSAggrCDTDataCSV(tstep, var_hgt, pars, start, end, aws_dir)
}
\arguments{
\item{tstep}{the time step of the data.}

\item{var_hgt}{the variable code and observation height, form  <var code>_<height>.}

\item{pars}{parameters.}

\item{start}{start time.}

\item{end}{end time.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
CSV object
}
\description{
Get aggregated data in CDT format for download.
}
