% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downAWSAggrOneVarCSV}
\alias{downAWSAggrOneVarCSV}
\title{Get aggregated data.}
\usage{
downAWSAggrOneVarCSV(tstep, net_aws, var_hgt, start, end, aws_dir)
}
\arguments{
\item{tstep}{the time step of the data.}

\item{net_aws}{the network code and AWS ID, form <network code>_<AWS ID>. AWS network code, 1: campbell, 2: adcon}

\item{var_hgt}{the variable code and observation height, form  <var code>_<height>.}

\item{start}{start date.}

\item{end}{end date.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
CSV format object
}
\description{
Get aggregated data displayed on the chart for download.
}
