% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downAWSMinDataCSV}
\alias{downAWSMinDataCSV}
\title{Get minutes data.}
\usage{
downAWSMinDataCSV(net_aws, var_hgt, start, end, aws_dir)
}
\arguments{
\item{net_aws}{the network code and AWS ID, form <network code>_<AWS ID>. 
AWS network code, 1: campbell, 2: adcon}

\item{var_hgt}{the variable code and observation height, form  <var code>_<height>.}

\item{start}{start time.}

\item{end}{end time.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
JSON format object
}
\description{
Get minutes data for download.
}
