% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downAWSStatusTable}
\alias{downAWSStatusTable}
\title{Download AWS status.}
\usage{
downAWSStatusTable(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/GMet_AWS_v2"}
}
\value{
CSV object
}
\description{
Download AWS status table.
}
