% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downHourlyMSLP}
\alias{downHourlyMSLP}
\title{Compute hourly mean sea level pressure.}
\usage{
downHourlyMSLP(time, aws_dir)
}
\arguments{
\item{time}{target date.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
CSV object
}
\description{
Compute hourly mean sea level pressure
}
