% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downTableAggrDataSelCSV}
\alias{downTableAggrDataSelCSV}
\title{Get aggregated data.}
\usage{
downTableAggrDataSelCSV(tstep, net_aws, var_hgt, pars, start, end, aws_dir)
}
\arguments{
\item{tstep}{the time step of the data.}

\item{net_aws}{a vector of the network code and AWS ID, form <network code>_<AWS ID>. AWS network code, 1: campbell, 2: adcon}

\item{var_hgt}{the variable code and observation height, form  <var code>_<height>.}

\item{pars}{parameters.}

\item{start}{start time.}

\item{end}{end time.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
CSV format object
}
\description{
Get aggregated data to download for multiple AWS.
}
