% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{getWindHeight}
\alias{getWindHeight}
\title{Get AWS Wind heights.}
\usage{
getWindHeight(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Get AWS Wind heights for dropdown select .
}
