% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plotmap_mslp.R
\name{mapHourlyMSLP}
\alias{mapHourlyMSLP}
\title{Compute hourly mean sea level pressure.}
\usage{
mapHourlyMSLP(time, aws_dir)
}
\arguments{
\item{time}{the time to display in the format "YYYY-MM-DD-HH"}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Compute hourly mean sea level pressure data to display on map.
}
