% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plotmap_aws.R
\name{mapMinAWSData}
\alias{mapMinAWSData}
\title{Get AWS 1 hour spatial data.}
\usage{
mapMinAWSData(time, aws_dir)
}
\arguments{
\item{time}{the time to display in the format "YYYY-MM-DD-HH-MM".}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Get AWS 1 hour spatial data to display on map.
}
