% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rain_accumulation.R
\name{mapRainAccumul}
\alias{mapRainAccumul}
\title{Compute precipitation accumulation.}
\usage{
mapRainAccumul(tstep, time, accumul, aws_dir)
}
\arguments{
\item{tstep}{time basis to accumulate the data.}

\item{time}{target date.}

\item{accumul}{accumulation duration.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Compute precipitation accumulation for spatial display.
}
