% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{readCoords}
\alias{readCoords}
\title{Read AWS metadata.}
\usage{
readCoords(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Read AWS coordinates and parameters.
}
