% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plotchart_aws.R
\name{tableAggrAWSData}
\alias{tableAggrAWSData}
\title{Get aggregated data.}
\usage{
tableAggrAWSData(tstep, net_aws, start, end, aws_dir)
}
\arguments{
\item{tstep}{the time step of the data.}

\item{net_aws}{the network code and AWS ID, form <network code>_<AWS ID>. 
AWS network code, 1: campbell, 2: adcon}

\item{start}{start date.}

\item{end}{end date.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Get aggregated data to display on table.
}
