% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{readCoordsWind}
\alias{readCoordsWind}
\title{Get AWS Wind coordinates.}
\usage{
readCoordsWind(height, aws_dir)
}
\arguments{
\item{height}{wind speed and direction heights above ground, format "<speedHeight>_<directionHeight>".}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/ZMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Get AWS Wind coordinates to display on map.
}
